+++ expense-tracker-backend/src/routes/TransactionRoutes.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import TransactionControllers from '../controllers/TransactionControllers.js';
+import RecurringTransactionController from '../controllers/RecurringTransactionController.js';
 import { authenticateUser } from '../middleware/auth.js';
 
 const router = express.Router();
@@ -11,9 +12,9 @@ router.put('/updateTransaction', authenticateUser, TransactionControllers.update
 router.delete('/deleteTransaction', authenticateUser, TransactionControllers.deleteTransaction);
 
 // Recurring transaction routes
-router.post('/createRecurring', authenticateUser, TransactionControllers.createRecurringTransaction);
-router.get('/getRecurring', authenticateUser, TransactionControllers.getRecurringTransactions);
-router.put('/updateRecurring', authenticateUser, TransactionControllers.updateRecurringTransaction);
-router.delete('/deleteRecurring', authenticateUser, TransactionControllers.deleteRecurringTransaction);
+router.post('/createRecurring', authenticateUser, RecurringTransactionController.createRecurringTransaction);
+router.get('/getRecurring', authenticateUser, RecurringTransactionController.getRecurringTransactions);
+router.put('/updateRecurring', authenticateUser, RecurringTransactionController.updateRecurringTransaction);
+router.delete('/deleteRecurring', authenticateUser, RecurringTransactionController.deleteRecurringTransaction);
 
 export default router;
